requirements.txt
openai==1.91.0
google-api-python-client==2.174.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
python-dotenv==1.1.1